ПР14

+++ controllers/cards.js
@@ -1,8 +1,10 @@
+const BadRequestError = require('../errors/BadRequestError');
+const NotFoundError = require('../errors/NotFoundError');
+const ForbiddenError = require('../errors/ForbiddenError');
+
 const Card = require('../models/card');
 
-const { ERROR_CODE, ERROR_NOT_FOUND, ERROR_SERVER } = require('../utils/constants');
-
-const createCard = (req, res) => {
+const createCard = (req, res, next) => {
   const { _id } = req.user;
   const { name, link } = req.body;
 
@@ -12,102 +14,97 @@ const createCard = (req, res) => {
     })
     .catch((error) => {
       if (error.name === 'ValidationError') {
-        return res.status(ERROR_CODE).send({
-          message: 'Переданы некорректные данные при создании карточки.',
-        });
+        next(
+          new BadRequestError('Переданы некорректные данные при создании карточки.'),
+        );
+        return;
       }
-      return res
-        .status(ERROR_SERVER)
-        .send({ message: 'На сервере произошла ошибка' });
+      next(error);
     });
 };
 
-const getCards = (req, res) => {
+const getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.send({ data: cards }))
-    .catch(() => {
-      res.status(ERROR_SERVER).send({ message: 'На сервере произошла ошибка' });
-    });
+    .catch(next);
 };
 
-const deleteCard = (req, res) => {
+const deleteCard = (req, res, next) => {
   const { cardId } = req.params;
-  Card.findByIdAndRemove(cardId)
+  Card.findById(cardId)
     .then((card) => {
       if (!card) {
-        res
-          .status(ERROR_NOT_FOUND)
-          .send({ message: ' Карточка с указанным _id не найдена.' });
-      } else {
-        res.send({ data: card });
+        throw new NotFoundError('Карточка с указанным _id не найдена.');
       }
+      const owner = card.owner.toString();
+      if (req.user._id !== owner) {
+        throw new ForbiddenError('Чужие карточки удалить нельзя!');
+      }
+      return Card.findByIdAndRemove(cardId)
+        .then((deletedCard) => {
+          res.send({ card: deletedCard });
+        });
     })
     .catch((error) => {
       if (error.name === 'CastError') {
-        res.status(ERROR_CODE).send({
-          message: 'Переданы некорректные данные',
-        });
+        next(new BadRequestError('Переданы некорректные данные.'));
       } else {
-        res.status(ERROR_SERVER).send({ message: 'На сервере произошла ошибка' });
+        next(error);
       }
     });
 };
 
-const likeCard = (req, res) => {
+const likeCard = (req, res, next) => {
   const owner = req.user._id;
   const { cardId } = req.params;
 
   Card.findByIdAndUpdate(
     cardId,
     { $addToSet: { likes: owner } },
-    { new: true, runValidators: true },
+    { new: true },
   )
     .then((card) => {
       if (!card) {
-        res
-          .status(ERROR_NOT_FOUND)
-          .send({ message: 'Передан несуществующий _id карточки.' });
+        next(new NotFoundError('Карточка с указанным _id не найдена.'));
       } else {
         res.send({ data: card });
       }
     })
     .catch((error) => {
       if (error.name === 'CastError') {
-        res.status(ERROR_CODE).send({
-          message: 'Переданы некорректные данные',
-        });
-      } else {
-        res.status(ERROR_SERVER).send({ message: 'На сервере произошла ошибка' });
+        next(
+          new BadRequestError('Переданы некорректные данные.'),
+        );
+        return;
       }
+      next(error);
     });
 };
 
-const dislikeCard = (req, res) => {
+const dislikeCard = (req, res, next) => {
   const owner = req.user._id;
   const { cardId } = req.params;
 
   Card.findByIdAndUpdate(
     cardId,
     { $pull: { likes: owner } },
-    { new: true, runValidators: true },
+    { new: true },
   )
     .then((card) => {
       if (!card) {
-        res
-          .status(ERROR_NOT_FOUND)
-          .send({ message: 'Передан несуществующий _id карточки.' });
+        next(new NotFoundError('Карточка с указанным _id не найдена.'));
       } else {
         res.send({ data: card });
       }
     })
     .catch((error) => {
       if (error.name === 'CastError') {
-        res.status(ERROR_CODE).send({
-          message: 'Переданы некорректные данные',
-        });
-      } else {
-        res.status(ERROR_SERVER).send({ message: 'На сервере произошла ошибка' });
+        next(
+          new BadRequestError('Переданы некорректные данные.'),
+        );
+        return;
       }
+      next(error);
     });
 };
